8.1 Image Digits Code & data/Code/mysite/requirements.txt
absl-py==0.15.0
asgiref==3.4.1
astunparse==1.6.3
boto3==1.20.39
botocore==1.23.39
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
clang==5.0
cycler==0.11.0
Django==3.2
django-storages==1.12.3
flatbuffers==1.12
fonttools==4.28.5
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.1.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.10.1
jmespath==0.10.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
Pillow==8.3.0
protobuf==3.19.3
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
s3transfer==0.5.0
six==1.15.0
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.6.0
tensorflow-estimator==2.7.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.7.0
